+++ public/index.php
@@ -630,7 +630,7 @@ if ($result_content_cnt === 0) {
 </nav>
 
 <!-- Bootstrap -->
-<script src="./js/jquery-3.3.1.min.js"></script>
+<script src="./js/jquery-3.7.1.min.js"></script>
 <script src="./js/popper.min.js"></script>
 <script src="./js/bootstrap.min.js"></script>
 <!-- Drawer -->